flytt data class BeregnetAndel inn i OrdinærBarnetrygdUtil da den kun brukes der
